plmd/cmd_channel: debug

+++ core/pldm/cmd_channel/cmd_channel_tcp.c
@@ -184,6 +184,9 @@ int send_packet(struct cmd_channel *channel, struct cmd_packet *packet) {
 }*/
 
 int setup_server_socket() {
+    static pthread_mutex_t remove_socket_mutex = PTHREAD_MUTEX_INITIALIZER;
+    static bool socket_removed = false;
+    
     int sockfd = socket(AF_UNIX, SOCK_STREAM, 0);
     if (sockfd < 0) {
         platform_printf("socket"NEWLINE);
@@ -195,14 +198,20 @@ int setup_server_socket() {
     addr.sun_family = AF_UNIX;
     strncpy(addr.sun_path, CMD_CHANNEL_SOCKET_PATH, sizeof(addr.sun_path) - 1);
 
-    if (access(CMD_CHANNEL_SOCKET_PATH, F_OK) == 0) {
+    pthread_mutex_lock(&remove_socket_mutex);
+
+    if (!socket_removed && access(CMD_CHANNEL_SOCKET_PATH, F_OK) == 0) {
         platform_printf("Removing existing socket file" NEWLINE);
         if (unlink(CMD_CHANNEL_SOCKET_PATH) < 0) {
             platform_printf("unlink" NEWLINE);
             close(sockfd);
+            pthread_mutex_unlock(&remove_socket_mutex);
             return CMD_CHANNEL_SOCKET_ERROR;
         }
+        socket_removed = true;
     }
+
+    pthread_mutex_unlock(&remove_socket_mutex);
 
     if (bind(sockfd, (struct sockaddr *)&addr, sizeof(addr)) < 0) {
         platform_printf("bind"NEWLINE);
